LamaIndex/fastapi_lamaindex/requirements.txt

requests
fastapi
snowflake-connector-python
python-dotenv
